Update slack spotify script to take in name and webhook as params.
This is such a goofy script and a goofy language.
Anyways, I think the UX on using this feauter has to be seamless
otherwise no one will use it. So... I think let's start by making
the script a blackbox. And then let's see if we can wrap this in a
bot or a button. The next version here is just goign to be a copy
and posteable curl comand that will get the script and run it in
the bkg. Don't even download it. That way pusching updates will
be easier.
curl "github raw url" | osascript << with parameters(?)
something like that could work.
Obviously it's easiest as a slack bot right?
/spotify share
anyways...